requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

loguru == 0.5.1

pytest == 6.0.1

regex == 2020.7.14

requests == 2.24.0

cachetools == 4.2.0

JPype1 == 1.3.0

PyYAML == 5.4.1
